requirements.txt
Flask==2.2.3
Flask_JWT_Extended==4.4.4
flask_sqlalchemy==3.0.5
python_bcrypt==0.3.2
SQLAlchemy==2.0.4
